requirements.txt
pillow
google-genai
huggingface-hub
datasets
python-dotenv
matplotlib
numpy
pandas
wordcloud
nltk
